refactor(xcross): extract device log stream
